requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
HiYaPyCo==0.4.16
idna==3.3
Jinja2==2.11.3
loguru==0.6.0
MarkupSafe==2.0.1
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
retrying==1.3.3
schedule==1.1.0
shellingham==1.4.0
six==1.16.0
tornado==6.1
typer==0.3.2
typer-cli==0.0.12
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
